+++ ft_printf/printf_rules.mk
@@ -12,8 +12,14 @@ $(PRINTF_DEPDIR)/%.dep: ;
 .PRECIOUS: $(PRINTF_DEPDIR)/%.dep
 
 $(PRINTF_FCLEAN): $(PRINTF_CLEAN)
+ifeq ($(shell [ -e $(PRINTF) ] && echo 1 || echo 0),1)
 	rm -rf $(PRINTF)
+endif
+ifeq ($(shell [ -e $(PRINTF_DEPDIR) ] && echo 1 || echo 0),1)
 	rm -rf $(PRINTF_DEPDIR)
+endif
 
 $(PRINTF_CLEAN):
+ifeq ($(shell [ -e $(PRINTF_OBJS_DIR) ] && echo 1 || echo 0),1)
 	rm -rf $(PRINTF_OBJS_DIR)
+endif
